Data processing function
Trying to implement the data processing function received via Bluetooth, some changes are needed for memory usage

+++ Hist_Band/Drivers/uart.h
@@ -19,6 +19,7 @@ data_process: When a flag is activated take the data received and define new thr
 
 #include <stdbool.h>
 #include <stdlib.h>
+#include <string.h>
 #include <stdio.h>
 #include <avr/io.h>
 
@@ -27,6 +28,6 @@ data_process: When a flag is activated take the data received and define new thr
 
 void UART_init();
 void UART_SendString(const char *str);
-uint8_t data_process(uint8_t *lower_threshold, uint8_t *upper_threshold);
+uint8_t data_process(int *lower_threshold, int *upper_threshold);
 
 #endif /* UART_H_ */
